Refactor Rater entry to be with RaterCommand
